QA model: Change model.fit( ) arguments in accrodance with LSTM reqs
LSTM requires inputs to be 3D matrix with shape (batch_size, timesteps, features).
The input predefined for context has the shape of (num_contexts, features, num_words) where num_words varies due to contexts length differences, while the input predefined for query has the shape of (num_questions, features, num_words) which is quite the same to that of context. Therefore, we must change the shape of the input in the fit() function
However, num_contexts and num_questions are different (while model.fit requires them to be the same), this is because each context has more than one query related to it. And this WILL BE HANDLED LATER